Adjusted material values
added missing sword guard items and adjusted recipes
fixed crafting recipes for the simple version of copper and bronze tools
added tooltip information to tool parts
adjusted material costs at the smithing anvil
adjusted netherite upgrade rendering

adjusted material stat values, this will be an ongoing process

+++ src/main/java/amorphia/alloygery/content/material/AlloygeryMaterialHelper.java
@@ -29,10 +29,10 @@ public class AlloygeryMaterialHelper
 		AlloygeryMaterial bindingMaterial = getBindingMaterial(compound);
 		AlloygeryMaterial upgradeMaterial = getUpgradeMaterial(compound);
 
-		final float base = headMaterial.head_durability * bindingMaterial.durability_multiplier;
-		final float upgrade = base * upgradeMaterial.durability_multiplier + upgradeMaterial.head_durability;
+		final int base = (int) (headMaterial.head_durability * bindingMaterial.durability_multiplier);
+		final int upgrade = (int) (base * upgradeMaterial.durability_multiplier + upgradeMaterial.head_durability);
 
-		return (int) upgrade;
+		return upgrade;
 	}
 
 	public static int getEnchantability(NbtCompound compound)
